+++ Sources/HIGThemesContract/HIGTheme+Defaults.swift
@@ -72,6 +72,7 @@ extension HIGTheme {
     public var codeBlock: any HIGCodeBlockTokens { HIGSystemCodeBlockTokens() }
     public var carousel: any HIGCarouselTokens { HIGSystemCarouselTokens() }
     public var mediaRow: any HIGMediaRowTokens { HIGSystemMediaRowTokens() }
+    public var hero: any HIGHeroTokens { HIGSystemHeroTokens() }
     public var photoPicker: any HIGPhotoPickerTokens { HIGSystemPhotoPickerTokens() }
     public var photoEditor: any HIGPhotoEditorTokens { HIGSystemPhotoEditorTokens() }
     public var longTextEditor: any HIGLongTextEditorTokens { HIGSystemLongTextEditorTokens() }