+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -70,6 +70,7 @@ public protocol HIGTheme: Sendable {
     var codeBlock: any HIGCodeBlockTokens { get }
     var carousel: any HIGCarouselTokens { get }
     var mediaRow: any HIGMediaRowTokens { get }
+    var hero: any HIGHeroTokens { get }
     var photoPicker: any HIGPhotoPickerTokens { get }
     var photoEditor: any HIGPhotoEditorTokens { get }
     var longTextEditor: any HIGLongTextEditorTokens { get }